#1195: fix simulation not working on IE 11.
Replace Array.from() with spread syntax.
Replace dynamic imports with require() and callback.
Add simEditButtons to list of divs to be shown in case of IE.
Add else block for SpeechSynthesis.cancel().

Note: Array.from() and dynamic imports (which return a Promise object) are not compatible with IE. Transpiled ts code was thus erroneous for use with IE.